Macro To Compare Sheets
A macro to compare different values between two sheets and highlight the difference.  The last part of the macro will write the number of different values at the end of the row.

Note this doesn't work when comparing case and CaSe as excel doesn't take the capitalization into account when comparing values.